requirements.txt
flask
openai
redis
requests
python-dotenv
